API_search_engine/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
sentence-transformers==2.6.1
scikit-learn==1.4.2
rank_bm25==0.2.2
faiss-cpu==1.8.0
qdrant-client==1.9.1
nltk==3.8.1
numpy==1.26.4
pydantic==2.7.1
python-dotenv==1.0.1